feature: multiple accounts (#19)
This adds the ability to have multiple accounts for a user. Transferring
between accounts is still not possible, however, this is the start of
support for having multiple accounts for customers. In addition to not
being able to transfer between your own accounts, you will not be able
to close accounts that have been opened. Even if this was done by
accident.

We have capped the number of accounts a customer can have to 2. This is
to prevent mass amounts of data flowing through our system. We felt like
in most cases, this will work. In the future, we might bump this number
to 3, however we do not expect it to go over that.

This was a requested feature to be brought back to Fun Banking, so here
we are :)

Co-authored-by: = <104793781+gomarcello@users.noreply.github.com>